Gen2: Adapter from HTTP GET to HTTP POST calls, this is a pre-req to implement authentication; auth implemented, but not yet working
